CDP #163 - Adding getDate helper function to Typesense utils
